refactoring anon and user limiter
